requirements.txt
flask==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Bcrypt-Flask==1.0.2
WTForms==3.1.2
email-validator==2.0.0
Authlib==1.3.0
requests-oauthlib==2.0.0
